Experiments finished :^)
Well, when deadline comes, git isn't as important anymore :-D

+++ scripts/metrics.py
@@ -3,10 +3,11 @@
 import argparse
 from pathlib import Path
 
-from skimage.metrics import structural_similarity 
+from skimage.metrics import structural_similarity
 from skimage.metrics import mean_squared_error
 from skimage.metrics import normalized_root_mse
 from skimage.metrics import peak_signal_noise_ratio
+import numpy as np
 
 parser = argparse.ArgumentParser(description="Read file and crop into given section")
 parser.add_argument("phantom", type=Path, help="Path to edf file to be windowed")
@@ -19,14 +20,14 @@ if __name__ == "__main__":
     reconstruction_path = args.reconstruction
     import edf
 
-    phantom = edf.readedf(phantom_path)
-    reconstruction = edf.readedf(reconstruction_path)
+    phantom = edf.readedf(phantom_path).astype(np.float64)
+    reconstruction = edf.readedf(reconstruction_path).astype(np.float64)
 
     MSE = mean_squared_error(phantom, reconstruction)
     NRMSE = normalized_root_mse(phantom, reconstruction)
-    # PSNR = peak_signal_noise_ratio(phantom, reconstruction, data_range=phantom.max() - phantom.min())
     PSNR = peak_signal_noise_ratio(phantom, reconstruction)
-    SSIM = structural_similarity(phantom, reconstruction, data_range=phantom.max() - phantom.min())
+    SSIM = structural_similarity(
+        phantom, reconstruction, data_range=phantom.max() - phantom.min()
+    )
 
-    # print(f"TEST: {MSE = }, {NRMSE = }, {PSNR = }, {SSIM = }")
     print(f"{MSE = :.3e}, {NRMSE = :.3e}, {PSNR = :4.3F}, {SSIM = :4.3F}")
